backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
sqlmodel==0.0.14
pymysql==1.1.0
cryptography==42.0.0
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic[email]==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
alembic==1.13.1
python-dotenv==1.0.0
aiofiles==23.2.1
pillow==10.2.0
httpx==0.26.0
